feat(bridge): 桥接自热重载 + 进程级一次性 hello
把单文件 1788 行桥接拆为稳定壳 index.js + impl/<version>/ 版本化实现：
- 壳每 2s 探测 package.json version + impl/<version>/index.js mtime，变化时
  dispose 旧实现 → import('?t=mtime') 破缓存 → 挂载新实现（query 直击实现文件，
  不经 export* 间接层避免 ESM 缓存命中；版本/版本号用 fs 读取规避 CJS require.cache）；
- hello 为进程级一次性握手：首次 apply 写 1 条，热重载/重装 skipHello 不重发，
  新版本由后续每条记录携带（pet 按记录逐条校验）；
- apply 幂等：DSH 进程内反复 apply（实测 ~2s/次）不重复建 mux/queue/监听，
  防 approval/question 帧散落多路连接丢失（'审批变事后'）；
- agent 快照重放：热重载后枚举旧实现存活 agent 恢复聚合/控制；
- Node24 runner 悬挂加 --test-force-exit；新增 12 个热重载行为用例
  （幂等/缓存破坏/agent 重放/hello 一次性）。
